requirements.txt
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
cachetools==5.3.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
colorama==0.4.6
cryptography==39.0.1
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
Django==3.2.18
django-appconf==1.0.5
django-debug-toolbar==3.2.4
django-extensions==3.2.1
django-imagekit==4.1.0
entrypoints==0.4
executing==1.2.0
Faker==13.3.5
fastjsonschema==2.16.2
fqdn==1.5.1
google-auth==2.16.1
google-auth-oauthlib==1.0.0
gspread==5.7.2
httplib2==0.21.0
idna==3.3
ipykernel==6.17.1
ipython==8.11.0
ipython-genutils==0.2.0
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.5.0
jupyter_client==7.4.7
jupyter_core==5.1.0
jupyter_server==2.0.1
jupyter_server_terminals==0.4.2
jupyterlab-pygments==0.2.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
mysql-connector-python==8.0.32
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.0
oauth2client==4.1.3
oauthlib==3.2.2
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pilkit==2.0
Pillow==9.5.0
platformdirs==2.5.4
prometheus-client==0.15.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.14.0
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==0.21.0
python-json-logger==2.0.4
pytz==2023.3
pywin32==305
pywinpty==2.0.9
PyYAML==6.0
pyzmq==24.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.4
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
tornado==6.2
traitlets==5.9.0
tzdata==2021.5
uri-template==1.2.0
urllib3==1.26.10
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
